Modified front-end to adhere to Model View Controller clean architecture.
