+++ frontend/interface_adapter/StopsData.tsx
@@ -1,0 +1,8 @@
+import LatLngLiteral = google.maps.LatLngLiteral;
+
+export type Stops = {
+    [index: string]: {
+        latitude: number;
+        longitude: number;
+    }
+}
