+++ frontend/styles/globals.css
@@ -37,7 +37,7 @@ a {
 }
 
 .map {
-  width: 80%;
+    width: 80%;
   height: 100vh;
 }
