+++ frontend/view/places.tsx
